test(memory): remove obsolete consolidation offset coverage
