+++ tests/agent/test_loop_consolidation_tokens.py
@@ -84,7 +84,7 @@ async def test_prompt_above_threshold_archives_until_next_user_boundary(tmp_path
     archive_end = loop.consolidator.archive_session.await_args.kwargs["archive_end"]
     archived_chunk = session.messages[:archive_end]
     assert [message["content"] for message in archived_chunk] == ["u1", "a1", "u2", "a2"]
-    assert session.last_consolidated == 4
+    assert session.last_archived == 4
 
 
 @pytest.mark.asyncio
@@ -123,7 +123,7 @@ async def test_consolidation_loops_until_target_met(tmp_path, monkeypatch) -> No
     )
 
     assert loop.consolidator.archive_session.await_count == 2
-    assert session.last_consolidated == 6
+    assert session.last_archived == 6
 
 
 @pytest.mark.asyncio
@@ -163,7 +163,7 @@ async def test_consolidation_continues_below_trigger_until_half_target(tmp_path,
     )
 
     assert loop.consolidator.archive_session.await_count == 2
-    assert session.last_consolidated == 6
+    assert session.last_archived == 6
 
 
 @pytest.mark.asyncio
